frontend/requirements.txt
asgiref==3.3.4
attrs==21.4.0
aws-cfn-bootstrap==2.0
awscli==2.9.19
awscrt==0.16.7
Babel==2.9.1
certifi==2021.10.8
cffi==1.14.5
chardet==4.0.0
chevron==0.13.1
cloud-init==22.2.2
colorama==0.4.4
configobj==5.0.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
dbus-python==1.2.18
distro==1.5.0
Django==3.2.3
djangorestframework==3.12.4
docutils==0.16
drf-yasg==1.20.0
ec2-hibinit-agent==1.0.2
gpg==1.15.1
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==0.10.0
jsonpatch==1.21
jsonpointer==2.0
jsonschema==4.5.1
libcomps==0.1.18
lockfile==0.12.2
MarkupSafe==2.1.1
netifaces==0.10.6
oauthlib==3.0.2
packaging==21.3
ply==3.11
prettytable==0.7.2
prompt-toolkit==3.0.24
pycparser==2.20
pyparsing==3.0.8
pyrsistent==0.18.1
pyserial==3.4
PySocks==1.7.1
python-daemon==2.3.0
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.3.1
release-notification==1.2
requests==2.25.1
rpm==4.16.1.3
ruamel.yaml==0.16.6
ruamel.yaml.clib==0.1.2
selinux==3.4
sepolicy==3.4
setools==4.4.1
six==1.16.0
sqlparse==0.4.1
support-info==1.0
swagger-spec-validator==2.7.4
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==1.26.9
wcwidth==0.2.5
